Bug fixes related to calculation of Neighborhood defoliation and site-level defoliation calculation.  Site defoliation is now the biomass-weighted average defoliation proportion across all present cohorts when using Biomass Succession.  An overload of the method weights site defoliation evenly among all cohorts (not biomass-weighted) if biomass is not available.
Modified Beta distributions to accept values of 0 for the alpha and beta parameters.  This allows the specification of distributions that always return either 0 or 1.

Added susceptibility class 4 for species.  These are intended to represent species that provide a protective effect to reduce defoliation levels.  Class 4 species  are treated the same as class 3 for susceptibility calculations, but the proportion of Class 4 biomass is used as a linear reduction in defoliation proportion.

git-svn-id: http://Marc-PC/svn/Full@1522 85f15e0e-5137-b144-a863-98bd542365a3